+++ src/mock/point.js
@@ -1,4 +1,4 @@
-import { getRandomInteger } from '../utils.js';
+import { getRandomInteger } from '../utils/common.js';
 import { TYPES, EXTRA_OFFERS, basePrice, destinationId } from './consts.js';
 import dayjs from 'dayjs';
 
@@ -25,4 +25,3 @@ export const generatePoint = () => {
     type: TYPES[getRandomInteger(0, TYPES.length - 1)],
   };
 };
-
